Hacky way to load weather data into watch app
On friday, we did some horrible things to get "real" data showing up on the
watch (the weather phrase and icon). Basically, Justin and I don't really know
Objective-C, so we were looking for a shortcut spikey way to hack in the data.

So the current behavior is:
1.) When the Tropos iPhone app refreshes, two different text files are written -
one with the phrase text, and the other with the name of the weather icon that
should be used.
2.) When the watch app is loaded, it calls back into the parent app using
openParentApplication.
3.) The iPhone app defines handleWatchKitExtensionRequest, which loads the text
files that were written, and passes a dictionary with the image name and phrase
text back to the apple watch extension via the openParentApplication callback.
4.) The watch extension updates the watch app's image and phrase text using the
contents of the dictionary.

I think the next step here is to extract a class in the iPhone app that performs
the weather request on the fly, to return a "fresh" phrase and image name,
instead of relying on the iPhone app to write to silly text files. =)